+++ seafoodveggies/src/main/java/com/store/seafoodveggies/service/AddressService.java
@@ -7,7 +7,7 @@ public interface AddressService {
     void saveAddress(Address address);
     Address updateAddress(Address address);
     void deleteAddress(Address address);
-    Address updateAddress(int userId, Long addressId, Address updatedAddress);
-    void deleteAddress(int userId, Long addressId);
+    Address updateAddress(Long userId, Long addressId, Address updatedAddress);
+    void deleteAddress(Long userId, Long addressId);
 
 }
