+++ seafoodveggies/src/main/java/com/store/seafoodveggies/service/UserService.java
@@ -6,6 +6,6 @@ import com.store.seafoodveggies.entity.User;
 public interface UserService {
      public User saveUser(User user);
      User findByUserName(String userName);
-     void deleteUser(int userId);
-     User updateUser(int id, User updatedUser);
+     void deleteUser(Long userId);
+     User updateUser(Long id, User updatedUser);
 }
